// src/app/api/ui/enums/lot-state.enum.ts
export enum LotStateEnum {
  Publication = 1,
  FilingApplications = 2,
  ReviewApplications = 3,
  Withdrawn = 4,
  AuctionWaiting = 5,
  NotTookPlace = 6,
  AuctionInProgress = 7,
  DeterminingWinner = 8,
  DealSigning = 9,
  DealComplete = 10,
  Suspended = 11,
  RetradingWaiting = 12,
  RetradingInProgress = 13,
  ReviewFirstPartApplications = 14,
  ReviewSecondPartApplications = 15,
  DealRejected = 16,
  ReviewingWaiting = 17,
  FinalProposalsReviewing = 18,
  FilingPreliminaryApplications = 19,
  RefinementDocuments = 20,
  OpenAccess = 21,
  EvaluationApplication = 22,
  ReadyToPublish = 23,
  SpecificationProposalsDiscussion = 24,
  SpecificationProposalsDiscussionWaiting = 25,
  FinalProposalsReviewWaiting = 26,
  DeterminingWinnerWaiting = 27,
  SpecificationDiscussion = 28,
  Qualification = 29,
  RetradingReview = 30,
  Archive = 31,
  NoApplications = 32,
  GeneralDiscussionAndRefinementDocuments = 96,
  GeneralFilingApplications = 97,
  Active = 98,
  GeneralTrading = 99,
  Forming = 100,
}